+++ src/redturtle/volto/restapi/serializer/summary.py
@@ -13,6 +13,11 @@ from redturtle.volto.interfaces import IRedturtleVoltoLayer
 from zope.component import adapter
 from zope.interface import implementer
 from zope.interface import Interface
+from plone.app.uuid.utils import uuidToCatalogBrain
+from plone.restapi.interfaces import IObjectPrimaryFieldTarget
+from zope.component import queryMultiAdapter
+from plone.restapi.serializer.utils import uid_to_url
+
 
 import re
 
@@ -89,19 +94,13 @@ class DefaultJSONSummarySerializer(BaseSerializer):
             # it isn't an internal link, so we can return it
             return value
         path = replace_link_variables_by_paths(context=self.context, url=value)
-        match = RESOLVEUID_RE.match(path)
-        if match:
-            uid, suffix = match.groups()
-            return uuidToURL(uid)
-        else:
-            portal = api.portal.get()
-            try:
-                ref_obj = portal.restrictedTraverse(path, None)
-                if ref_obj:
-                    return ref_obj.absolute_url()
-            except Exception:
-                return ""
-        return ""
+
+        url = uid_to_url(path)
+
+        if url == path:
+            # something wrong with the path, maybe a missing object?
+            return ""
+        return url
 
     def __call__(self, force_all_metadata=False):
         if force_all_metadata: